% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{An example data used for test PoDA}
\format{
A  list with elements
\describe{
  \item{count}{count matrix, 500 taxa by 100 samples}
  \item{interest}{the phenotype of interest, denoted as \code{u}}
  \item{confoun}{two confounders, denoted as \code{z1} and \code{z2}}
}
}
\usage{
example_data
}
\description{
A dataset containing "count", "interest", "confoun"
}
\keyword{datasets}
