% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{The imputation method proposed by Zhou et al.}
\usage{
imputation(Y, Z, formula)
}
\arguments{
\item{Y}{matrix or data frame; observed abundance matrix. Row: taxa; column: samples. NAs are not expected in observed abundance matrix.}

\item{Z}{matrix or data frame of covariates; The rows of \code{Z} correspond to the columns of \code{Y}. NAs are allowed.}

\item{formula}{character. For example: \code{formula = '~x1*x2+x3'}.}
}
\value{
The observed abundance matrix after value imputation.
}
\description{
The imputation method proposed by Zhou et al.
}
