% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PODA.R
\name{PODA}
\alias{PODA}
\title{Post-selection differential abundance analysis of microbiome compositional data}
\usage{
PODA(
  Y = Y,
  u = u,
  confoun = NULL,
  rep_time = 50,
  tau = 1,
  fdrnomial = 0.05,
  fdrmethod = "BH",
  lib_cut = 1000,
  pre_cut = 0.2,
  nCore = 9
)
}
\arguments{
\item{Y}{matrix; observed abundance matrix. Row: taxa; column: samples. NAs are not expected in observed abundance matrix.}

\item{u}{vector; the phenotype of interest.}

\item{confoun}{matrix; the other phenotypes to be adjusted.}

\item{rep_time}{a positive integer; the number of repeated information splitting (rep_time > 1)}

\item{tau}{a positive integer; it indicates the amount of information contained in each part after the data has undergone information-splitting.}

\item{fdrnomial}{a real value between 0 and 1; the nominal level of FDR control}

\item{fdrmethod}{character; p-value adjusting approach.}

\item{lib_cut}{a non-negative real value; samples with less than lib.cut read counts are excluded.}

\item{pre_cut}{a real value between 0 and 1; taxa with prevalence (percentage of nonzeros) less than prev.cut are excluded.}

\item{nCore}{a positive integer; the number of cores used for computing}
}
\value{
Return a list:
\itemize{
\item rej  vector; the index of differential abundant taxa identified by PoDA
\item rejtaxon.name  vector; the names of differential abundant taxa identified by PoDA
\item rej_atleast vector; the names of taxa that are considered differentially abundant taxa in at least one of the multiple information splitting replications.
\item all_freq  data frame; for all taxa, the number of times each was considered a differential abundant taxa across multiple information splitting replications.
\item rej_freq  data frame; for those taxa identified by PoDA as differential abundant taxa, the number of times each was considered a differential abundant taxa across multiple information splitting replications
}
}
\description{
Post-selection differential abundance analysis of microbiome compositional data
}
